tests/wpt/tests/encoding-detection/ja-ISO-2022-JP.tentative.html

<!doctype html>
<title>ja ISO-2022-JP</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>$B$3$l$OJ8;z<B83$G$9!#(B</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1252", 'Expected windows-1252 due to ISO-2022-JP no longer getting detected');
    }, "Check detection result");
      done();
};
</script>
